pyproject.toml
[project]
name = "job-search-web-app"
version = "0.1.0"
description = "A FastAPI-based web application for job search"
authors = [
    {name = "orin277",email = "bogush.vasyar@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "google-genai (>=1.25.0,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

packages = [
    { include = "app" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

